removes redundant mock data for knative  and update specs
