+++ frontend/packages/dev-console/src/utils/application-utils.ts
@@ -27,7 +27,7 @@ import { ClusterServiceVersionModel } from '@console/operator-lifecycle-manager'
 import {
   ServiceModel as KnativeServiceModel,
   RouteModel as KnativeRouteModel,
-} from '@console/knative-plugin';
+} from '@console/knative-plugin/src/models';
 import { isDynamicEventResourceKind } from '@console/knative-plugin/src/utils/fetch-dynamic-eventsources-utils';
 import { checkAccess } from '@console/internal/components/utils';
 import { getOperatorBackedServiceKindMap } from '@console/shared';
