+++ frontend/packages/knative-plugin/src/components/overview/__tests__/EventSinkServicesOverviewList.spec.tsx
@@ -5,15 +5,12 @@ import { referenceForModel, PodKind } from '@console/internal/module/k8s';
 import { PodControllerOverviewItem } from '@console/shared';
 import { PodsOverview } from '@console/internal/components/overview/pods-overview';
 import {
-  sampleEventSourceApiServer,
-  sampleEventSourceCamel,
-} from '@console/dev-console/src/components/topology/__tests__/topology-knative-test-data';
-import {
   ResourceLink,
   ExternalLink,
   SidebarSectionHeading,
 } from '@console/internal/components/utils';
-import { ServiceModel } from '../../../models';
+import { getEventSourceResponse } from '../../../topology/__tests__/topology-knative-test-data';
+import { ServiceModel, EventSourceApiServerModel, EventSourceCamelModel } from '../../../models';
 import EventSinkServicesOverviewList from '../EventSinkServicesOverviewList';
 
 describe('EventSinkServicesOverviewList', () => {
@@ -66,13 +63,19 @@ describe('EventSinkServicesOverviewList', () => {
   ];
 
   it('should show error info if no sink present or sink,kind is incorrect', () => {
-    const wrapper = shallow(<EventSinkServicesOverviewList obj={sampleEventSourceCamel.data[0]} />);
+    const mockData = _.omit(
+      _.cloneDeep(getEventSourceResponse(EventSourceCamelModel).data[0]),
+      'spec',
+    );
+    const wrapper = shallow(<EventSinkServicesOverviewList obj={mockData} />);
     expect(wrapper.find('span').text()).toBe('No services found for this resource.');
   });
 
   it('should have ResourceLink with proper kind', () => {
     const wrapper = shallow(
-      <EventSinkServicesOverviewList obj={sampleEventSourceApiServer.data[0]} />,
+      <EventSinkServicesOverviewList
+        obj={getEventSourceResponse(EventSourceApiServerModel).data[0]}
+      />,
     );
     const findResourceLink = wrapper.find(ResourceLink);
     expect(findResourceLink).toHaveLength(1);
@@ -81,20 +84,28 @@ describe('EventSinkServicesOverviewList', () => {
 
   it('should have ExternaLink when sinkUri is present', () => {
     const wrapper = shallow(
-      <EventSinkServicesOverviewList obj={sampleEventSourceApiServer.data[0]} />,
+      <EventSinkServicesOverviewList
+        obj={getEventSourceResponse(EventSourceApiServerModel).data[0]}
+      />,
     );
     expect(wrapper.find(ExternalLink)).toHaveLength(1);
   });
 
   it('should not have ExternalLink when no sinkUri is present', () => {
-    const mockEventSourceDataNoURI = _.omit(sampleEventSourceApiServer.data[0], 'status');
+    const mockEventSourceDataNoURI = _.omit(
+      getEventSourceResponse(EventSourceApiServerModel).data[0],
+      'status',
+    );
     const wrapper = shallow(<EventSinkServicesOverviewList obj={mockEventSourceDataNoURI} />);
     expect(wrapper.find(ExternalLink)).toHaveLength(0);
   });
 
   it('should show Deployment if present', () => {
     const wrapper = shallow(
-      <EventSinkServicesOverviewList obj={sampleEventSourceApiServer.data[0]} current={current} />,
+      <EventSinkServicesOverviewList
+        obj={getEventSourceResponse(EventSourceApiServerModel).data[0]}
+        current={current}
+      />,
     );
     const findResourceLink = wrapper.find(ResourceLink);
     const findSidebarSectionHeading = wrapper.find(SidebarSectionHeading);
@@ -107,20 +118,23 @@ describe('EventSinkServicesOverviewList', () => {
   it('should show pods if present', () => {
     const wrapper = shallow(
       <EventSinkServicesOverviewList
-        obj={sampleEventSourceApiServer.data[0]}
+        obj={getEventSourceResponse(EventSourceApiServerModel).data[0]}
         current={current}
         pods={pods}
       />,
     );
     expect(wrapper.find(PodsOverview)).toHaveLength(1);
     expect(wrapper.find(PodsOverview).props().allPodsLink).toEqual(
-      '/search/ns/testproject1?kind=Pod&q=sources.knative.dev%2FapiServerSource%3Dtestevents',
+      '/search/ns/testproject3?kind=Pod&q=sources.knative.dev%2FapiServerSource%3Doverlayimage',
     );
   });
 
   it('should not show pods if not present', () => {
     const wrapper = shallow(
-      <EventSinkServicesOverviewList obj={sampleEventSourceApiServer.data[0]} current={current} />,
+      <EventSinkServicesOverviewList
+        obj={getEventSourceResponse(EventSourceApiServerModel).data[0]}
+        current={current}
+      />,
     );
     expect(wrapper.find(PodsOverview)).toHaveLength(0);
   });
